Reorganizing tree tests
